support jdk 1.4
git-svn-id: http://svn.aei.mpg.de/eScience/GridSphere/trunk@4895 9c99c85f-4d0c-0410-8460-a9a1c48a3a7f

+++ src/org/gridlab/gridsphere/portlet/impl/SportletContext.java
@@ -267,11 +267,11 @@ public class SportletContext implements PortletContext {
     }
 
     public int getMajorVersion() {
-        return Integer.valueOf(GridSphereConfig.getProperty("gridsphere.portletapi.majorversion"));
+        return Integer.valueOf(GridSphereConfig.getProperty("gridsphere.portletapi.majorversion")).intValue();
     }
 
     public int getMinorVersion() {
-        return Integer.valueOf(GridSphereConfig.getProperty("gridsphere.portletapi.minorversion"));
+        return Integer.valueOf(GridSphereConfig.getProperty("gridsphere.portletapi.minorversion")).intValue();
     }
 
     /**
